Simplify SmartContractResults. Breaking change.
