+++ src/smartcontracts/smartContractResults.ts
@@ -5,25 +5,12 @@ import { GasLimit, GasPrice } from "../networkParams";
 import { Nonce } from "../nonce";
 import { TransactionHash } from "../transaction";
 import { ArgSerializer } from "./argSerializer";
-import { EndpointDefinition, TypedValue } from "./typesystem";
-import { ReturnCode } from "./returnCode";
-import { Result } from "./result";
 
 export class SmartContractResults {
     private readonly items: SmartContractResultItem[] = [];
-    private readonly immediate: TypedResult = new TypedResult();
-    private readonly resultingCalls: TypedResult[] = [];
 
     constructor(items: SmartContractResultItem[]) {
         this.items = items;
-
-        if (this.items.length > 0) {
-            let immediateResult = this.findImmediateResult();
-            if (immediateResult) {
-                this.immediate = immediateResult;
-            }
-            this.resultingCalls = this.findResultingCalls();
-        }
     }
 
     static empty(): SmartContractResults {
@@ -35,35 +22,9 @@ export class SmartContractResults {
         return new SmartContractResults(items);
     }
 
-    private findImmediateResult(): TypedResult | undefined {
-        let immediateItem = this.items.filter(item => isImmediateResult(item))[0];
-        if (immediateItem) {
-            return new TypedResult(immediateItem);
-        }
-        return undefined;
+    getAll(): SmartContractResultItem[] {
+        return this.items;
     }
-
-    private findResultingCalls(): TypedResult[] {
-        let otherItems = this.items.filter(item => !isImmediateResult(item));
-        let resultingCalls = otherItems.map(item => new TypedResult(item));
-        return resultingCalls;
-    }
-
-    getImmediate(): TypedResult {
-        return this.immediate;
-    }
-
-    getResultingCalls(): TypedResult[] {
-        return this.resultingCalls;
-    }
-
-    getAllResults(): TypedResult[] {
-        return this.items.map(item => new TypedResult(item));
-    }
-}
-
-function isImmediateResult(item: SmartContractResultItem): boolean {
-    return item.nonce.valueOf() != 0;
 }
 
 export class SmartContractResultItem {
@@ -78,7 +39,6 @@ export class SmartContractResultItem {
     gasLimit: GasLimit = new GasLimit(0);
     gasPrice: GasPrice = new GasPrice(0);
     callType: number = 0;
-    returnMessage: string = "";
 
     static fromHttpResponse(response: {
         hash: string,
@@ -107,68 +67,11 @@ export class SmartContractResultItem {
         item.gasLimit = new GasLimit(response.gasLimit);
         item.gasPrice = new GasPrice(response.gasPrice);
         item.callType = response.callType;
-        item.returnMessage = response.returnMessage;
 
         return item;
     }
 
     getDataTokens(): Buffer[] {
         return new ArgSerializer().stringToBuffers(this.data);
-    }
-}
-
-export class TypedResult extends SmartContractResultItem implements Result.IResult {
-    /**
-    * If available, will provide typed output arguments (with typed values).
-    */
-    endpointDefinition?: EndpointDefinition;
-
-    constructor(init?: Partial<SmartContractResultItem>) {
-        super();
-        Object.assign(this, init);
-    }
-
-    assertSuccess() {
-        Result.assertSuccess(this);
-    }
-
-    isSuccess(): boolean {
-        return this.getReturnCode().isSuccess();
-    }
-
-    getReturnCode(): ReturnCode {
-        let tokens = this.getDataTokens();
-        if (tokens.length < 2) {
-            return ReturnCode.None;
-        }
-        let returnCodeToken = tokens[1];
-        return ReturnCode.fromBuffer(returnCodeToken);
-    }
-
-    outputUntyped(): Buffer[] {
-        this.assertSuccess();
-
-        // Skip the first 2 SCRs (eg. the @6f6b from @6f6b@2b).
-        return this.getDataTokens().slice(2);
-    }
-
-    setEndpointDefinition(endpointDefinition: EndpointDefinition) {
-        this.endpointDefinition = endpointDefinition;
-    }
-
-    getEndpointDefinition(): EndpointDefinition | undefined {
-        return this.endpointDefinition;
-    }
-
-    getReturnMessage(): string {
-        return this.returnMessage;
-    }
-
-    outputTyped(): TypedValue[] {
-        return Result.outputTyped(this);
-    }
-
-    unpackOutput(): any {
-        return Result.unpackOutput(this);
     }
 }
